Microservices/targeting-service/requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.5